// src/app/types/homePageType.ts
export interface PricingSliceType {
	game: string;
	service: string;
	region: string;
	rankStart: string;
	rankStartPoint: string;
	rankTo: string;
	rankToPoint: string;
	appearOfflineFlag: string;
	agentLockFlag: string;
	soloFlag: string;
	duoFlag: string;
	streamFlag: string;
	priorityCompleteFlag: string;
	coupons: string;
	price: string;
	priceAfterDiscount: string;
}